Code cleanup

+++ lib/features/payment/domain/use_case/payment_use_case.dart
@@ -1,8 +1,8 @@
-import 'package:rent_hub/core/utils/snakbar/snackbar_utils.dart';
+import 'package:rent_hub/core/utils/snackbar_utils.dart';
 import 'package:rent_hub/features/payment/service/payment_service.dart';
 
 class PaymentUseCase {
-  void call(int amount) {
+  void call(double amount) {
     try {
       // This is a dummy method to simulate payment
       PaymentService().doPayment(amount);
